Suppressed build/run of various mpi demo drivers.
This is to get around problems arising from the fact that
the default linear solver doesn't switch to  superlu_dist any more.

Disabled the following codes when we're configuring without hypre and/or
trilinos:

- mpi/solvers/driven_cavity
- mpi/solvers/airy_cantilever
- mpi/solvers/fsi_channel_with_leaflet

Also changed order of solvers in direct_solvers.cc test to make
sure the test pass when certain solvers (such as superlu_dist) are
disabled. Also adjusted the number of tests in the validate.sh script
to handle this gracefully.

Much of this can be reversed if/when we get superlu_dist to work again.