Temporay commit to get Anthony involved in incorporating Hassan's gmsh code.
Currently processing the gmsh input takes prohibitively long and the code
that translate gmsh to oomph-lib is in a right state that cannot be
commited as is. The rest seems better and the newly added self-test is
clean and passes. It just takes for bloody ever to run (for a tiny example!)